Extract service boundary CLI
